Adjusted app view for small devices
Application now looks as supposed to on small screens.

I am not really fully satisfied with application design, but I don't
want to work on it. I created this application just to gain basic
knowledge about vue.js, so I can use it for a project on one of the
classes at the university. App served its purpose and I am not going to
develop it any further.